refactor routes - remove legacy governance pages
