+++ packages/yoroi-extension/app/Routes.js
@@ -45,17 +45,9 @@ import { CatalystRegistrationContextProvider } from './UI/features/catalyst-regi
 // $FlowIgnore: suppressing this error
 import { DappCenterContextProvider } from './UI/features/dapp-center/module/DappCenterContextProvider';
 // $FlowIgnore: suppressing this error
-import GovernanceDelegationFormPage from './UI/pages/Governance/GovernanceDelegationFormPage';
-// $FlowIgnore: suppressing this error
-import GovernanceRevampStatusPage from './UI/pages/Governance-Revamp/GovernanceRevampStatusPage';
-// $FlowIgnore: suppressing this error
-import GovernanceOptionsPage from './UI/pages/Governance-Revamp/GovernanceOptionsPage';
+import GovernanceOptionsPage from './UI/pages/Governance/GovernanceOptionsPage';
 // $FlowIgnore: suppressing this error
 import GovernanceStatusPage from './UI/pages/Governance/GovernanceStatusPage';
-// $FlowIgnore: suppressing this error
-import GovernanceTransactionFailedPage from './UI/pages/Governance/GovernanceTransactionFailedPage';
-// $FlowIgnore: suppressing this error
-import GovernanceTransactionSubmittedPage from './UI/pages/Governance/GovernanceTransactionSubmittedPage';
 // $FlowIgnore: suppressing this error
 import PortfolioDappsPage from './UI/pages/portfolio/PortfolioDappsPage';
 // $FlowIgnore: suppressing this error
@@ -254,14 +246,8 @@ export const YoroiRoutes = (stores: StoresMap): Node => {
           </Route>
 
           <Route element={<GovernanceSubpages stores={stores} />}>
-            <Route path={ROUTES.Governance.ROOT} element={<GovernanceStatusPage stores={stores} />} />
-            <Route path={ROUTES.Governance.DELEGATE} element={<GovernanceDelegationFormPage stores={stores} />} />
-            <Route path={ROUTES.Governance.SUBMITTED} element={<GovernanceTransactionSubmittedPage stores={stores} />} />
-            <Route path={ROUTES.Governance.FAIL} element={<GovernanceTransactionFailedPage stores={stores} />} />
-          </Route>
-          <Route element={<GovernanceRevampSubpages stores={stores} />}>
-            <Route path={ROUTES.GOVERNANCE_REVAMP.ROOT} element={<GovernanceRevampStatusPage stores={stores} />} />
-            <Route path={ROUTES.GOVERNANCE_REVAMP.OPTIONS} element={<GovernanceOptionsPage stores={stores} />} />
+            <Route path={ROUTES.GOVERNANCE.ROOT} element={<GovernanceStatusPage stores={stores} />} />
+            <Route path={ROUTES.GOVERNANCE.OPTIONS} element={<GovernanceOptionsPage stores={stores} />} />
           </Route>
           <Route element={<PortfolioSubpages stores={stores} />}>
             <Route path={ROUTES.PORTFOLIO.ROOT} element={<PortfolioPage stores={stores} />} />
@@ -383,35 +369,7 @@ const CatalystRegistrationSubpages = ({ stores }) => (
   </CatalystRegistrationContextProvider>
 );
 
-// NEW UI - TODO: to be refactred
 const GovernanceSubpages = ({ stores }) => {
-  const { unitOfAccount } = stores.profile;
-  const currentWalletInfo = createCurrrentWalletInfo(stores);
-  const { delegationTransaction } = stores.substores.ada;
-  const delegationTxResult = delegationTransaction.createDelegationTx.result;
-  const delegationTxError = delegationTransaction.createDelegationTx.error;
-
-  return (
-    <CurrencyProvider currency={unitOfAccount.currency || 'USD'}>
-      <GovernanceContextProvider
-        currentWallet={currentWalletInfo}
-        createDrepDelegationTransaction={request => stores.delegation.createDrepDelegationTransaction(request)}
-        signDelegationTransaction={request => stores.substores.ada.delegationTransaction.signTransaction(request)}
-        txDelegationResult={delegationTxResult}
-        txDelegationError={delegationTxError}
-        tokenInfo={stores.tokenInfoStore.tokenInfo}
-        triggerBuySellAdaDialog={() => stores.uiDialogs.open({ dialog: BuySellDialog })}
-        getCurrentPrice={stores.coinPriceStore.getCurrentPrice}
-      >
-        <Suspense fallback={null}>
-          <Outlet />
-        </Suspense>
-      </GovernanceContextProvider>
-    </CurrencyProvider>
-  );
-};
-
-const GovernanceRevampSubpages = ({ stores }) => {
   const { unitOfAccount } = stores.profile;
   const currentWalletInfo = createCurrrentWalletInfo(stores);
   const { delegationTransaction } = stores.substores.ada;
