--- packages/yoroi-extension/app/UI/pages/Governance-Revamp/GovernanceRevampStatusPage.tsx
@@ -1,17 +1,0 @@
-import GovernanceLayout from './layout';
-import { GovernanceStatusRevamp } from '../../features/governace/useCases/GovernanceStatusRevamp/GovernanceStatusRevamp';
-
-type Props = {
-  stores: any;
-  children?: any;
-};
-
-const GovernanceStatusPage = (props: Props): any => {
-  return (
-    <GovernanceLayout {...props}>
-      <GovernanceStatusRevamp />
-    </GovernanceLayout>
-  );
-};
-
-export default GovernanceStatusPage;
