--- packages/yoroi-extension/app/UI/pages/Governance/GovernanceDelegationFormPage.tsx
@@ -1,17 +1,0 @@
-import { DelagationForm } from '../../features/governace/useCases/DelagationForm/DelagationForm';
-import GovernanceLayout from './layout';
-
-type Props = {
-  stores: any;
-  children?: any;
-};
-
-const GovernanceDelegationFormPage = (props: Props): any => {
-  return (
-    <GovernanceLayout {...props}>
-      <DelagationForm />
-    </GovernanceLayout>
-  );
-};
-
-export default GovernanceDelegationFormPage;
