+++ packages/yoroi-extension/app/routes-config.js
@@ -83,15 +83,10 @@ export const ROUTES = {
     // $FlowIgnore
     ERROR: '/swap-revamp' + PAGE_ERROR_SUBROUTE,
   },
-  Governance: {
+  GOVERNANCE: {
     ROOT: '/governance',
-    DELEGATE: '/governance/delagation',
-    SUBMITTED: '/governance/submitted',
+    OPTIONS: '/governance/options',
     FAIL: '/governance/failed',
-  },
-  GOVERNANCE_REVAMP: {
-    ROOT: '/governance-status',
-    OPTIONS: '/governance-status/options',
   },
   PORTFOLIO: {
     ROOT: '/portfolio',
